bcti-ai/ai_api/requirements.txt
fastapi
uvicorn[standard]
requests
numpy
sqlalchemy
psycopg2-binary
PyJWT